ADDED: Prepend '.' to the Python search path for _interactive_ sessions
Similar to Python, when Prolog is used interactively we add '' (or '.')
to ``sys.path`` as _first_ element.   Prints an informational message
when this happens.

Note that adding . for non-interative sessions is undesirable for
security reasons: unaware users of an application may trigger the
execution of code that happens to be in the current directory.